finance chatbot/requirements.txt
streamlit
pandas
numpy
plotly
pymongo
python-dotenv
prophet
cmdstanpy